+++ src/extension.ts
@@ -17,6 +17,7 @@ import statusBarItem, {
 import hover from "./utils/components/hover";
 import getDailySummary from "./utils/github/getDailySummary";
 import {
+  backendURL,
   EXTENSION_ID,
   GITHUB_AUTH_PROVIDER_ID,
   SCOPES,
@@ -33,6 +34,7 @@ import selectCommandHandler from "./utils/commands/select";
 import debugLogger from "./utils/vscode/debugLogger";
 import checkIfUserStarred from "./utils/github/checkIfUserStarred";
 import getMostRelevantJiraTicket from "./utils/jira/getMostRelevantJiraTicket";
+import getAssignedJiraTickets from "./utils/jira/getAssignedJiraTickets";
 import { WatermelonAuthenticationProvider } from "./auth";
 
 // repo information
@@ -167,6 +169,11 @@ export async function activate(context: vscode.ExtensionContext) {
         },
       });
 
+    // call our API to get assigned Jira tickets here
+    const jiraTickets = await getAssignedJiraTickets({
+      userEmail: session.account.label,
+    });
+
       let dailySummary = await getDailySummary({
         octokit,
         owner: owner || "",
@@ -176,7 +183,7 @@ export async function activate(context: vscode.ExtensionContext) {
       debugLogger(`dailySummary: ${JSON.stringify(dailySummary)}`);
       provider.sendMessage({
         command: "dailySummary",
-        data: dailySummary,
+        data: { dailySummary, jiraTickets },
       });
       if (startLine === undefined && endLine === undefined) {
         if (!arrayOfSHAs.length) {