+++ src/utils/vscode/getInitialHTML.ts
@@ -121,6 +121,8 @@ export default function getInitialHTML(
            <button class='run-watermelon btn btn-primary' type='button'>Get Code Context</button>
         </div>
         <h2>Daily Summary</h2>
+        <div id="jiraHolder"></div>
+        <br/>
         <div id="dailySummary"></div>
         You will find all your issues and ToDos here.